Remove "Load Salt Key" Preference.
The "Salt Key" preference itself will be clickable to load
the salt key from a file. It will not be directly editable,
since it does not make sense to edit a base64 encoded 512-byte
key by hand.

Minor fix: When restarting the "Settings" activity, the default
number of PBKDF2 iterations, when set by the user, is correctly
displayed.