Integrate Picker with "Salt Key" Preference
Integrated a "picker" with the "Salt Key" click handler.
The preference is enabled when the "Unlock Salt Key" checkbox
is checked, and clicking on it pulls up a picker to select
a text file containing the key. It will do its business,
and subsequently the "Unlock Salt Key" preference checkbox
will be unchecked and the "Salt Key" preference disabled.

At present, this does nothing. The URI obtained from the picker
needs to be processed.