Created the ability to Generate or Load a Salt Key
The salt key generation/loading was moved from the Preferences screen
to its own activity, since tying down this process to the Preferences
APIs encumbered it too much. Instead, the salt key appears on the
Preferences screen as a read-only attribute.
The new "SaltKeyActions" activity as well as the Preferences activity
can be navigated to using action buttons.
Loading the salt key happens through designated Content Providers, which,
at the moment, are the default ones - "Recent", "My Drive", and "Downloads".

Outstanding issues:
===================

a) No functionality for custom attributes has been implemented yet
b) A navigation drawer is needed, to navigate to children activities
   as well as to support the options to import/export preferences.
c) More Content Providers for loading the salt key need to be
   investigated.