Implemented Import/Export Settings
gobbledygook now features the ability to import/export preferences.

Importing is done via opening an intent to "get content", which will
allow the user to pick a file manager installed on the system to be able
to select a file from "external" storage.

Exporting is done in the form a JSON formatted file to a folder
called "gobbledygook" in "Documents" folder on the "external" storage,
allowing the user to consume this file elsewhere (such as in the browser
add-on, which still needs to implement this ability).

TODO:
=====

a) Fix up the saving custom preferences code
b) UI enhancements on the "About" page
c) Browser plugin catch-up w.r.t. domain name resolution