Code Refactoring for Supporting Product Flavors
Refactored all the code to support the two product flavors
"Gobbledygook" and "Krunch!". The two flavors share the same code, and
override the methods needed from the main Activity ("Yggdrasil"), which
is an abstract class, and is extended in the flavors.

More refactoring is in store as more features are added to "Krunch!"