A New Paradigm
This commit features a redesign of how the Gobbledygook workflow is
going to be. For seamless integration, a WebView has been incorporated
into the application so that users open webpages right into the in-built
browser, and generate passwords from there itself. The "home" fragment
is now not the main workhorse of the system, but rather the WebView, and
the workhorse is an ever-present dialog attached to the WebView.

This commit also features a transition from the Holo theme to the
Material design theme. There will be further changes down the line.

Armed with our new namespace, we will pierce the heavens!