More Materialism; Sane Versioning
Replaced the clunky ListView adapter implementation of the Navigation
Drawer with the Material Design NavigationView introduced in the design
support library.

Replaced the custom "About" page and its clunky ListView adapter
implementation with a PreferenceFragment that seamlessly integrates with
Material Design.

The "Salt Key" fragment is not a standalone activity anymore, but a
dialog that is accessible from the Settings page directly on clicking
the "Salt Key" preference. This should streamline user workflow.

Changed versioning scheme in build.gradle to create:
    * android:versionCode as YYYYMMDD
    * android:versionName as 1.YYYYMMDD.gitCommitSHA
This assumes (reasonably) that version granularity of less than a day
will not be needed. The versionName should serve to uniquely identify a
version, and the versionCode part (minor version) is monotonically
increasing as well.

Bug: The settings items in the Navigation Drawer do not remain clicked.