Added "No Special Characters" as an Option
This is to prepare for a transition to Z85 encoding, which allows for
special characters and therefore yields "stronger" passwords. At the
moment, the setting is a no-op, but the pipeline to respect it has been
laid, and an override for this setting will be saved in the schema.
Z85-integration will come at a later stage, when either:
https://github.com/bitwiseshiftleft/sjcl/pull/312
is approved for official Z85-support in sjcl, or the sjcl submodule is
changed to work off of the Z85 branch.