Browser Extension with Z85 Encoding
The browser extension now supports generating passwords with Z85
encoding. Both the Z85 and Base64 modules have unit tests, and the
extension is ready for daily use!

The Android module has the "No special characters" option, but still
needs support for Z85 encoding, as well as needs the "Workhorsefragment"
dialog to be cleaned up (collapse "advanced" options", etc). This will
be completed at a later stage when the mechanism for performing Z85
encoding in Java has been finalized.