Unit-Tests for Android; BugFix for Browser/Android.
Added junit tests for Android.

* While extending the tests for the browser extension, it was discovered
that the "truncation" logic was faulty (it truncated to one more
character than asked for). Also, the browser unit-tests were broken
after refactoring out variables in the file "env.js", indicating I never
ran the tests again - a heinous crime, which I have atoned for, and hope
to never commit again!
* While adding unit tests for Android, a number of bugs were discovered!
Most of them pertain to not using the "equals" comparison for objects,
as well as not overridding the "equals" method if equality tests on
custom Objects must be performed. Hooray for unit tests!
* Travis will automatically run the android tests on each checkin.