Android: Hide "Advanced Options" by Default
This is achieved by encapsulating the "advanced" elements into a
LinearLayout which is "GONE" by default. A checkbox allows toggling the
visibility, and the "animateLayoutChanges" attribute for the parent view
handles the animation natively.
Also, bugfix: the generate button handler was being passed the Button
view, but assumed it was the entire "Workhorse" view. This has been
corrected by passing in the entire view. Sadly, this sort of thing must
be "integration tested", as there is no automation for this yet.