CoordinatorLayout/WebView Enhancements, Icon Theme
When scrolling the WebView, it is desired to maximize screen real
estate. The App Bar therefore needed an overhaul to play well with the
CoordinatorLayout, as did the FloatingActionButton. This was made
possible by moving the toolbar to each respective fragment and then
setting the "support" ActionBar from the ToolBar, as well as the
ActionBarDrawerToggle. This allows the ActionBar and the
FloatingActionButton to disappear when scrolling down (view scrolls up)
and reappear when scrolling up. The WebView has been modified to
implement the NestedScrollingChild interface, to send nested scroll
events to the CoordinatorLayout.

Moving the ToolBar to each fragment requires special care with those of
type PreferenceFragment, and those were nested inside the FrameLayout of
a container fragment, which otherwise had the ToolBar. This is because
setting preferences from an XML resource does not play well with
inflating a custom layout for that view.

The icon set has been made consistent, borrowing from the Material Theme.