Browser Integration
Allow the app to receive intents from browsers via the "Share" option -
this allows browsers to "send" Gobbledygook the URL to generate a
password. Using this method, users are free to use browsers of their own
choice while using Gobbledygook purely as a password generator. The
built-in WebView is still available, but this mechanism may be a
necessary alternative for workflows which the internal WebView
cannot handle (and we do not wish to embark on the folly of making the
internal WebView a full-featured browser).

Also, minor bugfix for the "Krunch!" flavor - it was launching the
"Workhorse" fragment when asked to launch the "Home" fragment.